style
explicitly cast sizeof(char) to int since it never flow over INT_MAX

+++ cli.c
@@ -63,7 +63,7 @@ int main(int argc, char *argv[])
         printf("[CLIENT] ");
         fgets(buf, svr_buf_len, stdin);
         buf[strcspn(buf, "\n")] = '\0';
-        iRet = send(sock, buf, svr_buf_len * sizeof(char), 0);
+        iRet = send(sock, buf, svr_buf_len * (int)sizeof(char), 0);
         if (iRet == SOCKET_ERROR)
         {
             fprintf(stderr, "send failed with error: %d\n", WSAGetLastError());
@@ -73,7 +73,7 @@ int main(int argc, char *argv[])
             exit(EXIT_FAILURE);
         }
 
-        iRet = recv(sock, buf, svr_buf_len * sizeof(char), 0);
+        iRet = recv(sock, buf, svr_buf_len * (int)sizeof(char), 0);
         buf[svr_buf_len - 1] = '\0';
         if (iRet == SOCKET_ERROR)
         {
@@ -89,7 +89,7 @@ int main(int argc, char *argv[])
     shutdown(sock, SD_SEND);
     do
     {
-        iRet = recv(sock, buf, svr_buf_len * sizeof(char), 0);
+        iRet = recv(sock, buf, svr_buf_len * (int)sizeof(char), 0);
     }
     while (iRet != 0);
